# openapi.json
{"openapi":"3.1.0","info":{"title":"WorkforceIQ API","version":"0.1.0"},"paths":{"/api/v1/auth/login":{"post":{"tags":["auth"],"summary":"Login","operationId":"login_api_v1_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Login"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/register":{"post":{"tags":["auth"],"summary":"Register","operationId":"register_api_v1_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Register"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/me":{"get":{"tags":["auth"],"summary":"Get Me","operationId":"get_me_api_v1_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/dashboard/kpis":{"get":{"tags":["dashboard"],"summary":"Get Kpis","operationId":"get_kpis_api_v1_dashboard_kpis_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/employees":{"get":{"tags":["employees"],"summary":"List Employees","operationId":"list_employees_api_v1_employees_get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EmployeeResponse"},"title":"Response List Employees Api V1 Employees Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["employees"],"summary":"Create Employee","operationId":"create_employee_api_v1_employees_post","security":[{"OAuth2PasswordBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/skills/dummy":{"get":{"tags":["skills"],"summary":"Dummy","operationId":"dummy_api_v1_skills_dummy_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/forecast/dummy":{"get":{"tags":["forecast"],"summary":"Dummy","operationId":"dummy_api_v1_forecast_dummy_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/gap-analysis/dummy":{"get":{"tags":["gap_analysis"],"summary":"Dummy","operationId":"dummy_api_v1_gap_analysis_dummy_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/learning/dummy":{"get":{"tags":["learning"],"summary":"Dummy","operationId":"dummy_api_v1_learning_dummy_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/readiness/dummy":{"get":{"tags":["readiness"],"summary":"Dummy","operationId":"dummy_api_v1_readiness_dummy_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/promotion/dummy":{"get":{"tags":["promotion"],"summary":"Dummy","operationId":"dummy_api_v1_promotion_dummy_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/attrition/dummy":{"get":{"tags":["attrition"],"summary":"Dummy","operationId":"dummy_api_v1_attrition_dummy_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/roi/dummy":{"get":{"tags":["roi"],"summary":"Dummy","operationId":"dummy_api_v1_roi_dummy_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/reports/dummy":{"get":{"tags":["reports"],"summary":"Dummy","operationId":"dummy_api_v1_reports_dummy_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/settings/dummy":{"get":{"tags":["settings"],"summary":"Dummy","operationId":"dummy_api_v1_settings_dummy_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/notifications/dummy":{"get":{"tags":["notifications"],"summary":"Dummy","operationId":"dummy_api_v1_notifications_dummy_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/search/dummy":{"get":{"tags":["search"],"summary":"Dummy","operationId":"dummy_api_v1_search_dummy_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/copilot/dummy":{"get":{"tags":["copilot"],"summary":"Dummy","operationId":"dummy_api_v1_copilot_dummy_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/export/dummy":{"get":{"tags":["export"],"summary":"Dummy","operationId":"dummy_api_v1_export_dummy_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"EmployeeCreate":{"properties":{"first_name":{"type":"string","title":"First Name"},"last_name":{"type":"string","title":"Last Name"},"email":{"type":"string","format":"email","title":"Email"},"role":{"type":"string","title":"Role"},"department_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Department Id"},"location":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Location"},"salary":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Salary"},"hire_date":{"type":"string","format":"date","title":"Hire Date"},"manager_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Manager Id"},"status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status","default":"active"},"is_remote":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Remote","default":false}},"type":"object","required":["first_name","last_name","email","role","hire_date"],"title":"EmployeeCreate"},"EmployeeResponse":{"properties":{"first_name":{"type":"string","title":"First Name"},"last_name":{"type":"string","title":"Last Name"},"email":{"type":"string","format":"email","title":"Email"},"role":{"type":"string","title":"Role"},"department_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Department Id"},"location":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Location"},"salary":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Salary"},"hire_date":{"type":"string","format":"date","title":"Hire Date"},"manager_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Manager Id"},"status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status","default":"active"},"is_remote":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Remote","default":false},"id":{"type":"integer","title":"Id"},"performance_score":{"type":"number","title":"Performance Score"},"attrition_risk_score":{"type":"number","title":"Attrition Risk Score"},"attrition_risk_level":{"type":"string","title":"Attrition Risk Level"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["first_name","last_name","email","role","hire_date","id","performance_score","attrition_risk_score","attrition_risk_level","created_at"],"title":"EmployeeResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Login":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"Login"},"Register":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"},"first_name":{"type":"string","title":"First Name"},"last_name":{"type":"string","title":"Last Name"},"role":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Role","default":"employee"}},"type":"object","required":["email","password","first_name","last_name"],"title":"Register"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"refresh_token":{"type":"string","title":"Refresh Token"},"token_type":{"type":"string","title":"Token Type"}},"type":"object","required":["access_token","refresh_token","token_type"],"title":"Token"},"UserResponse":{"properties":{"id":{"type":"integer","title":"Id"},"email":{"type":"string","format":"email","title":"Email"},"first_name":{"type":"string","title":"First Name"},"last_name":{"type":"string","title":"Last Name"},"role":{"type":"string","title":"Role"},"avatar_initials":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Avatar Initials"},"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"is_active":{"type":"boolean","title":"Is Active"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","email","first_name","last_name","role","is_active","created_at"],"title":"UserResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/v1/auth/login"}}}}}}